druppie/mcp-servers/module-foundry/requirements.txt
fastmcp>=2.0.0,<3.0.0
uvicorn>=0.24.0
pyyaml
pydantic>=2.0,<3.0
azure-ai-projects>=2.0.0
azure-identity>=1.15.0
